+++ auto_repair_saas/apps/tenants/views.py
@@ -1,0 +1,14 @@
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseRedirect
+from django.urls import reverse
+
+from ..utils.factories import ContactFactory, JobFactory
+
+
+@login_required
+def seed_data(request, *args, **kwargs):
+    for _ in range(0, 15):
+        JobFactory()
+    for _ in range(0, 5):
+        ContactFactory(contact_type='supplier')
+    return HttpResponseRedirect(reverse('dashboard'))
