+++ auto_repair_saas/templates/contacts/index.html
@@ -57,7 +57,7 @@
                                     <img class="rounded-circle"
                                          height="35"
                                          loading="lazy" alt="{{ client.name }}"
-                                         src="https://i.pravatar.cc/40?u={{ client.name }}">
+                                         src="https://avatars.dicebear.com/4.5/api/avataaars/{{ client.name }}.svg">
                                     <span class="fs-5 ps-2">{{ client.name }}</span>
                                     <div class="float-end">
                                         <div class="dropdown pt-1">
@@ -124,7 +124,7 @@
                                     <img class="rounded-circle"
                                          height="35"
                                          loading="lazy" alt="{{ supplier.name }}"
-                                         src="https://ui-avatars.com/api/?background=random&name={{ supplier.name }}&length=1">
+                                         src="https://avatars.dicebear.com/4.5/api/avataaars/{{ supplier.name }}.svg">
                                     <span class="fs-5 ps-2">{{ supplier.name }}</span>
                                     <div class="float-end">
                                         <div class="dropdown pt-1">
