+++ auto_repair_saas/templates/dashboard/index.html
@@ -43,7 +43,7 @@
                         <div class="row d-flex align-items-center mb-2">
                             <div class="col-4">
                                 <img
-                                        src="https://i.pravatar.cc/30?u{{ staff }}"
+                                        src="https://avatars.dicebear.com/4.5/api/avataaars/{{ staff }}.svg"
                                         class="rounded-circle me-2"
                                         height="25"
                                         alt="profile"
